# Optimize dfs in extract_model by dict