### Description
<!-- - Describe your changes. -->

Use a dictionary `output_to_index` to map an output name to the index of the corresponding node.

### Motivation and Context
<!-- - Why is this change required? What problem does it solve? -->
<!-- - If it fixes an open issue, please link to the issue here. -->

There are some facts:
1. `reachable` and `unreachable` are mutually exclusive sets. `reachable` union `unreachable` equals to all nodes.
2. In onnx, output name is unique. `len(nodes_to_search)` is either 0 or 1.
```
nodes_to_search = [
    index
    for index in unreachable
    if current_output_name in nodes[index].output
]
```
This step can be speed up by using a `dict` to map an output name to a node.
After introducing a dict, there's no reason to keep `unreachable`, all unreachable checks can be done by `reachable`

### Experiment Results

I did an experiment to extract a unet model (around 3.3GB, 4900 nodes).
```
extract_model(
    input_path=model_path,
    output_path=model_path + ".tmp",
    input_names=[node.name for node in model.graph.input],
    output_names=[node.name for node in model.graph.output],
)
```
The execution time of the `Extractor.extract_model` function is reduced from 7.24 seconds to 1.15 seconds.
The `Extractor.extract_model` function is no longer the slowest part of the extract model process.
```
check_model 3.31s
infer_shapes 6.15s
Extractor.extract_model 1.16s
onnx.save 5.60s
check_model 2.40s

real    0m20.376s
user    0m7.284s
sys     0m14.980s
```